requirements.txt
anthropic
Pillow
imagehash
requests